refactor: Like 도메인 DTO 패키지 변경
