+++ src/main/java/org/ktc2/cokaen/wouldyouin/like/application/HostLikeService.java
@@ -2,8 +2,8 @@ package org.ktc2.cokaen.wouldyouin.like.application;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.ktc2.cokaen.wouldyouin.like.api.HostLikeResponse;
-import org.ktc2.cokaen.wouldyouin.like.api.LikeRequest;
+import org.ktc2.cokaen.wouldyouin.like.application.dto.HostLikeResponse;
+import org.ktc2.cokaen.wouldyouin.like.application.dto.LikeRequest;
 import org.ktc2.cokaen.wouldyouin.like.persist.HostLikeRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
